html/infrastructure/urls/resolving-urls/resolve-urls.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Resolve URLs</title>
<base id="base" target="_self">
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" title="2.5.2 Resolving URLs" href="http://www.w3.org/html/wg/drafts/html/CR/infrastructure.html#resolving-urls">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  //resolve relative URLs
  var a1 = document.createElement("a"),
      base = document.getElementById("base"),
      base_url = document.location.href,
      host = document.location.host,
      root = base_url.replace(document.location.pathname, "/"),
      path = base_url.substr(0, base_url.lastIndexOf("/"));

  var testData = [
    "/", root,
    "/test.txt", root + "test.txt",
    ".", path + "/",
    "..", path.substr(0, path.lastIndexOf("/") + 1),
    "test.txt", path + "/test.txt",
    "./test.txt", path + "/test.txt",
    "../test.txt", path.substr(0, path.lastIndexOf("/") + 1) + "test.txt",
    "../aaa/test.txt", path.substr(0, path.lastIndexOf("/") + 1) + "aaa/test.txt",
    "\u3167/test.txt", path + "/%E3%85%A7/test.txt"
  ];

  for (var i = 0; i < testData.length; i += 2) {
    test(function () {
      a1.setAttribute("href", testData[i]);
      assert_greater_than(a1.href.indexOf(host), -1, "The resolved URL must have a hostname.");
      assert_equals(a1.href, testData[i + 1], "The resolved URL should be " + testData[i + 1] + ".");
    }, "The url '" + testData[i] + "' must be resolved to '" + testData[i + 1] + "'");
  }

  //resolve absolute URLs
  test(function () {
    a1.setAttribute("href", "http://www.example.com");
    assert_equals(a1.href, "http://www.example.com/", "The resolved URL should be 'http://www.example.com/'.");
  }, "The url 'http://www.example.com' must be resolved to 'http://www.exam;ple.com/'");
  
  test(function () {
    a1.setAttribute("href", "http://www.ex\u0061mple.com");
    assert_equals(a1.href, "http://www.example.com/", "The resolved URL should be 'http://www.example.com/'.");
  }, "The url 'http://www.ex\\u0061mple.com' must be resolved to 'http://www.example.com/'");

  //Base is an absolute URL
  base.href = "http://www.example.com/test/";
  testData = [
    "/", "http://www.example.com/",
    "/test.txt", "http://www.example.com/test.txt",
    ".", "http://www.example.com/test/",
    "..", "http://www.example.com/",
    "test.txt", "http://www.example.com/test/test.txt",
    "./test.txt", "http://www.example.com/test/test.txt",
    "../test.txt", "http://www.example.com/test.txt",
    "../aaa/test.txt", "http://www.example.com/aaa/test.txt",
    "../../test.txt", "http://www.example.com/test.txt",
    "\u3167/test.txt", "http://www.example.com/test/" + "%E3%85%A7/test.txt",
    "http://www.example2.com/", "http://www.example2.com/"
  ];

  for (var i = 0; i < testData.length; i += 2) {
    test(function () {
      a1.setAttribute("href", testData[i]);
      assert_equals(a1.href, testData[i + 1], "The resolved URL should be " + testData[i + 1] + ".");      
    }, "The url '" + testData[i].replace('\\', "") + "' must be resolved to '" + testData[i + 1] + "'");
  }
</script>
